+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/models/cart/CartImpl.java
@@ -54,6 +54,8 @@ public class CartImpl implements Cart, ModelBase {
 
     private com.commercetools.api.models.cart.TaxedPrice taxedShippingPrice;
 
+    private com.commercetools.api.models.cart.DiscountOnTotalPrice discountOnTotalPrice;
+
     private com.commercetools.api.models.cart.TaxMode taxMode;
 
     private com.commercetools.api.models.cart.RoundingMode taxRoundingMode;
@@ -123,6 +125,7 @@ public class CartImpl implements Cart, ModelBase {
             @JsonProperty("totalPrice") final com.commercetools.api.models.common.CentPrecisionMoney totalPrice,
             @JsonProperty("taxedPrice") final com.commercetools.api.models.cart.TaxedPrice taxedPrice,
             @JsonProperty("taxedShippingPrice") final com.commercetools.api.models.cart.TaxedPrice taxedShippingPrice,
+            @JsonProperty("discountOnTotalPrice") final com.commercetools.api.models.cart.DiscountOnTotalPrice discountOnTotalPrice,
             @JsonProperty("taxMode") final com.commercetools.api.models.cart.TaxMode taxMode,
             @JsonProperty("taxRoundingMode") final com.commercetools.api.models.cart.RoundingMode taxRoundingMode,
             @JsonProperty("taxCalculationMode") final com.commercetools.api.models.cart.TaxCalculationMode taxCalculationMode,
@@ -164,6 +167,7 @@ public class CartImpl implements Cart, ModelBase {
         this.totalPrice = totalPrice;
         this.taxedPrice = taxedPrice;
         this.taxedShippingPrice = taxedShippingPrice;
+        this.discountOnTotalPrice = discountOnTotalPrice;
         this.taxMode = taxMode;
         this.taxRoundingMode = taxRoundingMode;
         this.taxCalculationMode = taxCalculationMode;
@@ -310,7 +314,7 @@ public class CartImpl implements Cart, ModelBase {
     }
 
     /**
-     *  <p>Sum of the <code>totalPrice</code> field of all LineItems and CustomLineItems, and if available, the <code>price</code> field of ShippingInfo.</p>
+     *  <p>Sum of the <code>totalPrice</code> field of all LineItems and CustomLineItems, and if available, the <code>price</code> field of ShippingInfo. If a discount applies on <code>totalPrice</code>, this field holds the discounted value.</p>
      *  <p>Taxes are included if TaxRate <code>includedInPrice</code> is <code>true</code> for each price.</p>
      */
 
@@ -323,6 +327,7 @@ public class CartImpl implements Cart, ModelBase {
      *   <li>For a Cart with <code>Platform</code> TaxMode, it is automatically set when a shipping address is set.</li>
      *   <li>For a Cart with <code>External</code> TaxMode, it is automatically set when the external Tax Rate for all Line Items, Custom Line Items, and Shipping Methods in the Cart are set.</li>
      *  </ul>
+     *  <p>If a discount applies on <code>totalPrice</code>, this field holds the discounted values.</p>
      */
 
     public com.commercetools.api.models.cart.TaxedPrice getTaxedPrice() {
@@ -335,6 +340,14 @@ public class CartImpl implements Cart, ModelBase {
 
     public com.commercetools.api.models.cart.TaxedPrice getTaxedShippingPrice() {
         return this.taxedShippingPrice;
+    }
+
+    /**
+     *  <p>Discounts that apply on the Cart <code>totalPrice</code>.</p>
+     */
+
+    public com.commercetools.api.models.cart.DiscountOnTotalPrice getDiscountOnTotalPrice() {
+        return this.discountOnTotalPrice;
     }
 
     /**
@@ -622,6 +635,11 @@ public class CartImpl implements Cart, ModelBase {
         this.taxedShippingPrice = taxedShippingPrice;
     }
 
+    public void setDiscountOnTotalPrice(
+            final com.commercetools.api.models.cart.DiscountOnTotalPrice discountOnTotalPrice) {
+        this.discountOnTotalPrice = discountOnTotalPrice;
+    }
+
     public void setTaxMode(final com.commercetools.api.models.cart.TaxMode taxMode) {
         this.taxMode = taxMode;
     }
@@ -774,6 +792,7 @@ public class CartImpl implements Cart, ModelBase {
                 .append(totalPrice, that.totalPrice)
                 .append(taxedPrice, that.taxedPrice)
                 .append(taxedShippingPrice, that.taxedShippingPrice)
+                .append(discountOnTotalPrice, that.discountOnTotalPrice)
                 .append(taxMode, that.taxMode)
                 .append(taxRoundingMode, that.taxRoundingMode)
                 .append(taxCalculationMode, that.taxCalculationMode)
@@ -816,6 +835,7 @@ public class CartImpl implements Cart, ModelBase {
                 .append(totalPrice, that.totalPrice)
                 .append(taxedPrice, that.taxedPrice)
                 .append(taxedShippingPrice, that.taxedShippingPrice)
+                .append(discountOnTotalPrice, that.discountOnTotalPrice)
                 .append(taxMode, that.taxMode)
                 .append(taxRoundingMode, that.taxRoundingMode)
                 .append(taxCalculationMode, that.taxCalculationMode)
@@ -863,6 +883,7 @@ public class CartImpl implements Cart, ModelBase {
                 .append(totalPrice)
                 .append(taxedPrice)
                 .append(taxedShippingPrice)
+                .append(discountOnTotalPrice)
                 .append(taxMode)
                 .append(taxRoundingMode)
                 .append(taxCalculationMode)