+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/models/cart_discount/CartDiscountTarget.java
@@ -30,6 +30,7 @@ import io.vrap.rmf.base.client.utils.Generated;
         @JsonSubTypes.Type(value = com.commercetools.api.models.cart_discount.CartDiscountCustomLineItemsTargetImpl.class, name = CartDiscountCustomLineItemsTarget.CUSTOM_LINE_ITEMS),
         @JsonSubTypes.Type(value = com.commercetools.api.models.cart_discount.CartDiscountLineItemsTargetImpl.class, name = CartDiscountLineItemsTarget.LINE_ITEMS),
         @JsonSubTypes.Type(value = com.commercetools.api.models.cart_discount.CartDiscountShippingCostTargetImpl.class, name = CartDiscountShippingCostTarget.SHIPPING),
+        @JsonSubTypes.Type(value = com.commercetools.api.models.cart_discount.CartDiscountTotalPriceTargetImpl.class, name = CartDiscountTotalPriceTarget.TOTAL_PRICE),
         @JsonSubTypes.Type(value = com.commercetools.api.models.cart_discount.MultiBuyCustomLineItemsTargetImpl.class, name = MultiBuyCustomLineItemsTarget.MULTI_BUY_CUSTOM_LINE_ITEMS),
         @JsonSubTypes.Type(value = com.commercetools.api.models.cart_discount.MultiBuyLineItemsTargetImpl.class, name = MultiBuyLineItemsTarget.MULTI_BUY_LINE_ITEMS) })
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, include = JsonTypeInfo.As.EXISTING_PROPERTY, property = "type", defaultImpl = CartDiscountTargetImpl.class, visible = true)
@@ -67,6 +68,10 @@ public interface CartDiscountTarget {
             return com.commercetools.api.models.cart_discount.CartDiscountShippingCostTarget
                     .deepCopy((com.commercetools.api.models.cart_discount.CartDiscountShippingCostTarget) template);
         }
+        if (template instanceof com.commercetools.api.models.cart_discount.CartDiscountTotalPriceTarget) {
+            return com.commercetools.api.models.cart_discount.CartDiscountTotalPriceTarget
+                    .deepCopy((com.commercetools.api.models.cart_discount.CartDiscountTotalPriceTarget) template);
+        }
         if (template instanceof com.commercetools.api.models.cart_discount.MultiBuyCustomLineItemsTarget) {
             return com.commercetools.api.models.cart_discount.MultiBuyCustomLineItemsTarget
                     .deepCopy((com.commercetools.api.models.cart_discount.MultiBuyCustomLineItemsTarget) template);
@@ -101,6 +106,14 @@ public interface CartDiscountTarget {
      */
     public static com.commercetools.api.models.cart_discount.CartDiscountShippingCostTargetBuilder shippingBuilder() {
         return com.commercetools.api.models.cart_discount.CartDiscountShippingCostTargetBuilder.of();
+    }
+
+    /**
+     * builder for totalPrice subtype
+     * @return builder
+     */
+    public static com.commercetools.api.models.cart_discount.CartDiscountTotalPriceTargetBuilder totalPriceBuilder() {
+        return com.commercetools.api.models.cart_discount.CartDiscountTotalPriceTargetBuilder.of();
     }
 
     /**